Several fixes especially for FreeBSD
- Make sure events don't start before the config is updated
- Removed some debug code
- Ported ntp fix to rewrite
- Signed / unsigned fix in 433nano
- Memory allocation check in lirc protocol
- Improved ARP unit test
- Fixed event mail action for FreeBSD poll

+++ libs/pilight/hardware/433nano.c
@@ -263,9 +263,9 @@ int clock_gettime(int dummy, struct timespec *ct) {
 static void on_read(uv_fs_t *req) {
 	unsigned int len = req->result, pos = 0;
 	char *p = NULL;
-	if(len < 0) {
+	if((int)len < 0) {
 		logprintf(LOG_ERR, "read error: %s", uv_strerror(req->result));
-	} else if(len >= 0) {
+	} else if((int)len >= 0) {
 		memcpy(&timestamp.first, &timestamp.second, sizeof(struct timespec));
 		clock_gettime(CLOCK_MONOTONIC, &timestamp.second);
 		if((((double)timestamp.second.tv_sec + 1.0e-9*timestamp.second.tv_nsec) -
